+++ src/mixins/modules/recalculateMarkerAlignment.js
